Scaffolding for the AnalyserNode (time domain and frequency data)
It keeps a ring buffer of mono waveform data on the render thread.

Actual analysis will be performed on the render thread too, due to the
smoothing requirements. Otherwise we need to ship an infinite amount of
data to the control thread.

If the control thread requests time/frequency data, it sends a message
in a “rendezvous channel” (sync channel with size 0). This will be
handled in the next render quantum, and the result is pinged back.

This means only one request is handled per render quantum. This is
probably fine as the audio frame rate is much higher than a typical
visual frame rate.