RenderScope must not be Copy/Clone, pass reference to processors
When we allow processor to store state inside the RenderScope, it cannot
be copied obviously.

Now, we are handing out an (not mutable) reference to the RenderScope.
This prevent processors from changing the timestamp/frame/samplerate
value.

When we add the 'state' blobs, this will probably a
HashMap<String, Cell<Box<dyn Any>>> or something, and hide it between a
get/set method on the &RenderScope