AudioRenderQuantumChannel now derefs to [f32], not [f32; 128]
In preparation of 'variable render quantum size', it is important to
never explicitly expose the buffer size in the public API.

This has a small performance hit though because the new deref signature
forces more bound checks at runtime. It is in the order of a few percent
for some benchmarks.