Throw InvalidStateError when calling suspend while rendering has started
This is slightly different from the spec, where it says we should throw
if the suspend time is less than or equal to the current time. However,
this is more or less the same because rendering happens so fast the user
does not even know about the precise 'current time' anyway.