Add failing test case where NodeCollection would panic
This issue arises when a graph involves nodes connecting to audio
params.  E.g. constant src -> playback rate -> buffer src.

The playbackrate param is dropped alongside the buffer src if it has
ended, but this leaves an edge from the constant src in place.

To solve this we probably need a bit more refactoring, stay tuned