Change ScriptProcessorNode::onaudioprocess API - use owned buffer value
This is easier for the node bindings. We hook into the Drop call of the
event to ship the output buffer to the render thread.

TODO: I found a gnarly bug where the ScriptProcessorRenderer would be
cleared from the audio graph erroneously. AudioContextRegistration
should not be allowed to implement Clone because it has a very
meaningful Drop implementation. That's why I now resort to an `Arc` hack
to prevent the Drop method to run.